+++ .github/workflows/new-pr-to-main.yaml
@@ -18,11 +18,11 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  code_quality_check:
-    # Only run on PRs from the same repo
-    if: github.event.pull_request.head.repo.full_name == github.repository
-    uses: ./.github/workflows/code-quality-check.yml
-    secrets: inherit
+#  code_quality_check:
+#    # Only run on PRs from the same repo
+#    if: github.event.pull_request.head.repo.full_name == github.repository
+#    uses: ./.github/workflows/code-quality-check.yml
+#    secrets: inherit
   check_changeset:
     if: github.event.pull_request.head.repo.full_name == github.repository
     runs-on: ubuntu-latest
