+++ builders/changesetAction/tsconfig.json
@@ -1,4 +1,8 @@
 {
   "extends": "tsconfig/tsconfig.base.json",
+  "compilerOptions": {
+    "target": "es2015",
+    "moduleResolution": "node"
+  },
   "include": ["src"]
 }
